Material Font, SSN Lookangles, and Custom Sensors
Fixed an issue with IE and the material fonts. Still not loading perfectly on older
versions of IE, but now is at least useable.

SSN Lookangles had a bug in it where it would stop looking if a sensor didn't have any
passes at all. This is now fixed.

Custom sensors have been available since November, but now the current sensor data is put
in the URL to make it easier to remember what variables to change for a custom sensor.

In the future this should be migrated to a side menu, but is a temporary fix to display
telescopes/HAM radio FOVs.