North or South Option on Nominals
Nominal orbit creator checks last latitude to determine if it is heading north or south.
If it is going the wrong direction (compared to what the user selected) it then rotates
the right ascension (rasc) 180 degrees and continues until the latitude is found again.

This allows the user to launch any object from a list of launch sites in either a north or south
direction along the inclination of the original object.