Stereographic Map and Telescopes
Added a stereographic map to the bottom menu. This uses a modified version of the
propagate class of the lookangles function to calculate 50 points within the next orbit
and then displays them as links to the time that they occur. The end result is a ground
trace that highlights when things are in view and allows you to jump ahead to an event
of interest.

Also added telescope toggle to the custom sensors tab to make it easier for visitors to
quickly see what is above them. Should add auto geolocation next...